Renamed & moved lexanalysis package
